Cast dti to unsigned in two places to avoid -Wsign-compare

+++ libgadget/timebinmgr.c
@@ -180,7 +180,7 @@ dloga_from_dti(inttime_t dti)
         dti = -dti;
         sign = -1;
     }
-    if(dti > TIMEBASE) {
+    if((unsigned int) dti > TIMEBASE) {
         endrun(-1, "Requesting dti larger than TIMEBASE\n");
     }
     return Dloga * dti * sign;
